+++ client/src/features/relatedLists/SearchResultRelatedList.tsx
@@ -4,6 +4,7 @@ import { Link } from 'react-router-dom'
 import { Col, Row } from 'react-bootstrap'
 
 import StyledSearchLink from '../../styles/shared/SearchLink'
+import StyledHr from '../../styles/shared/Hr'
 import { formatHalLink } from '../../lib/parse/search/queryParser'
 import { IOrderedItems, ISearchResults } from '../../types/ISearchResults'
 import { getEstimates } from '../../lib/parse/search/searchResultParser'
@@ -11,6 +12,8 @@ import RecordLink from '../common/RecordLink'
 import { searchScope } from '../../config/searchTypes'
 import { getAllParamsFromHalLink } from '../../lib/parse/search/halLinkHelper'
 import { pushClientEvent } from '../../lib/pushClientEvent'
+import useResizeableWindow from '../../lib/hooks/useResizeableWindow'
+import theme from '../../styles/theme'
 
 interface IProps {
   url: string
@@ -54,6 +57,10 @@ const SearchResultRelatedList: React.FC<IProps> = ({
   data,
   title,
 }) => {
+  const [showHr, setShowHr] = useState<boolean>(
+    window.innerWidth < theme.breakpoints.md,
+  )
+
   const recordLinks = (orderedItems: Array<IOrderedItems>): any =>
     orderedItems.map((item, ind: number) => {
       const { id } = item
@@ -74,9 +81,12 @@ const SearchResultRelatedList: React.FC<IProps> = ({
     sort !== null ? `&${resultsEndpoint[0]}s=${sort}` : ''
   }`
 
+  useResizeableWindow(setShowHr)
+
   return (
     <React.Fragment>
       {recordLinks(orderedItems)}
+      {showHr && <StyledHr width="100%" className="mt-3" />}
       <StyledSearchLink className="row py-2 text-start">
         <Col xs={12} className="mt-1">
           <Link