+++ client/CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Changed mobile layout of landing page ([#428](https://github.com/project-lux/lux-frontend/issues/428)).
+- Changed mobile layout of search results accordions ([#440](https://github.com/project-lux/lux-frontend/issues/440)).
+- Changed mobile layout of facet accordions ([#441](https://github.com/project-lux/lux-frontend/issues/441)).
+- Changed mobile layout of related list accordions ([#443](https://github.com/project-lux/lux-frontend/issues/443)).
 
 ### Removed